feat: 音声合成機能のアーキテクチャ改善
- SpeechQueue: キュー管理機能を独立したクラスに分離
- AudioPlayer: 音声再生機能を独立したクラスに分離
- AudioSynthesizer: 音声合成機能を独立したクラスに分離
- 型定義を独立ファイル(types.ts)に整理
- ConfigManagerのキャッシュ処理を最適化（不要なディープコピー削除）
